Move asm.sh disassembler into the anal.sh ##refactor
