+++ libr/anal/meson.build
@@ -100,7 +100,8 @@ r_anal_sources = [
   join_paths('p','anal_s390_gnu.c'),
   join_paths('../asm/arch/s390/gnu/s390-dis.c'),
   join_paths('../asm/arch/s390/gnu/s390-opc.c'),
-  join_paths('p','anal_sh.c'),
+  join_paths('p/anal_sh.c'),
+  join_paths('../asm/arch/sh/gnu/sh-dis.c'),
   join_paths('p','anal_snes.c'),
   join_paths('p','anal_lm32.c'),
   join_paths('p','anal_sparc_cs.c'),
