+++ libr/anal/p/sh.mk
@@ -1,10 +1,10 @@
 OBJ_SH=anal_sh.o
+OBJ_SH+=../../asm/arch/sh/gnu/sh-dis.o
 
-STATIC_OBJ+=${OBJ_SH}
-TARGET_SH=anal_sh.${EXT_SO}
+STATIC_OBJ+=$(OBJ_SH)
+TARGET_SH=anal_sh.$(EXT_SO)
 
-ALL_TARGETS+=${TARGET_SH}
+ALL_TARGETS+=$(TARGET_SH)
 
-${TARGET_SH}: ${OBJ_SH}
-	${CC} $(call libname,anal_sh) ${LDFLAGS} \
-		${CFLAGS} -o anal_sh.${EXT_SO} ${OBJ_SH}
+$(TARGET_SH): $(OBJ_SH)
+	$(CC) $(call libname,anal_sh) $(LDFLAGS) $(CFLAGS) -o anal_sh.$(EXT_SO) $(OBJ_SH)
