+++ libr/asm/meson.build
@@ -24,7 +24,6 @@ r_asm_sources = [
   join_paths('p','asm_ppc_as.c'),
   join_paths('p','asm_ppc_cs.c'),
   join_paths('p','asm_riscv.c'),
-  join_paths('p','asm_sh.c'),
   join_paths('p','asm_sparc_gnu.c'),
   join_paths('p','asm_tms320.c'),
   join_paths('p','asm_x86_as.c'),
@@ -71,7 +70,6 @@ r_asm_sources = [
   join_paths('arch','or1k','or1k_disas.c'),
   #join_paths('arch','riscv','riscv-opc.c'),
   #join_paths('arch','riscv','riscv.c'),
-  join_paths('arch','sh','gnu','sh-dis.c'),
   join_paths('arch','sparc','gnu','sparc-dis.c'),
   join_paths('arch','sparc','gnu','sparc-opc.c'),
   join_paths('arch','tms320','c55x_plus','c55plus.c'),
